+++ src/app/stores/todo/todo-context.ts
@@ -1,5 +1,5 @@
 import { createContext } from "react"
-import { TTodoContextProps } from "../../types/states/types-todo"
+import { TTodoContextProps } from "../../types/stores/types-todo"
 
 export const TodoContext = createContext<TTodoContextProps>(
   {} as TTodoContextProps,
