+++ src/main.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import ReactDOM from "react-dom/client"
 import App from "./ui/pages/index"
 import "./index.css"
-import TodoProvider from "./app/states/todo/TodoProvider"
+import TodoProvider from "./app/stores/todo/TodoProvider"
 
 ReactDOM.createRoot(document.getElementById("root") as HTMLElement).render(
   <React.StrictMode>
